move every_country into utils
